Dashboards Mutation API: rename shared command helpers to shared.ts
