+++ public/app/features/dashboard-scene/mutation-api/commands/shared.ts
@@ -1,9 +1,10 @@
 /**
- * Shared helpers for reading dashboard-level settings off the DashboardScene.
+ * Helpers shared between mutation commands.
  *
- * Used by UPDATE_DASHBOARD_SETTINGS (to diff before/after) and GET_DASHBOARD_INFO
- * (to report current values), so the set of settings stays in sync between the
- * command that writes them and the command that reads them.
+ * readDashboardSettings/findCursorSyncBehavior are used by UPDATE_DASHBOARD_SETTINGS
+ * (to diff before/after) and GET_DASHBOARD_INFO (to report current values), so the
+ * set of settings stays in sync between the command that writes them and the one
+ * that reads them.
  */
 
 import { behaviors, sceneGraph } from '@grafana/scenes';
